requirement.txt
asgiref
astroid==2.4.2
autopep8==1.5
colorama==0.4.3
Django==3.0.6
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-tinymce==3.0.2
django-tinymce4-lite==1.8.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
isort==4.3.21
joblib==0.14.1
jsmin==2.2.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.2
pandas==1.0.3
Pillow==7.1.1
pycodestyle==2.5.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.2.post1
scipy==1.4.1
simplejson==3.17.0
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
toml==0.10.1
Unidecode==1.1.1
whitenoise==5.1.0
wrapt==1.12.1
yapf==0.29.0
django-rest-auth-0.9.5
djangorestframework-jwt-1.11.0